requirement.txt
fastapi 
uvicorn 
python-dotenv 
openai
google
langchain_community